[TOW-37] 테스트 뷰 제거 및 아이템 분리
* 탭 아이템 분리
* 테스트 뷰 삭제
