+++ TodayWod/Features/Common/CustomView/CustomTabView.swift
@@ -7,30 +7,6 @@
 
 import SwiftUI
 
-enum TabMenuItem: CaseIterable {
-    case home
-    case settings
-
-    var selectedImage: Image {
-        switch self {
-        case .home:
-            Images.icHomeActivate.swiftUIImage
-        case .settings:
-            Images.icAccountActivate.swiftUIImage
-        }
-    }
-
-    var deactivatedImage: Image {
-        switch self {
-        case .home:
-            Images.icHomeDisable.swiftUIImage
-        case .settings:
-            Images.icAccountDisable.swiftUIImage
-        }
-    }
-
-}
-
 struct CustomTabView: View {
 
     @Binding var selectedItem: TabMenuItem
